+++ pmj_server/src/lib.rs
@@ -13,5 +13,4 @@
 // 您應該已經收到一份 GNU Affero 通用公共授權條款副本。
 // 如果沒有，請參見 <https://www.gnu.org/licenses/>。
 
-pub mod gamemodes;
 pub mod shared;